+++ data/src/main/java/com/mashup/damgledamgle/repository/spec/response/NickNameRequest.kt
@@ -1,16 +1,15 @@
-package com.mashup.damgledamgle.repository.spec
+package com.mashup.damgledamgle.repository.spec.response
 
 import com.squareup.moshi.Json
 
 /**
- *  NotifyResponse.kt
+ *  NickNameRequest.kt
  *
- *  Created by Minji Jeong on 2022/08/02
+ *  Created by Minji Jeong on 2022/07/28
  *  Copyright © 2022 MashUp All rights reserved.
  */
 
-data class NotifyResponse(
+data class NickNameRequest(
     @Json(name = "nickname") val nickName: String,
-    @Json(name = "userNo") val userNo: Int,
     @Json(name = "notification") val notification: Boolean
 )
