+++ data/src/main/java/com/mashup/damgledamgle/repository/spec/response/NotifyResponse.kt
@@ -1,15 +1,16 @@
-package com.mashup.damgledamgle.repository.spec
+package com.mashup.damgledamgle.repository.spec.response
 
 import com.squareup.moshi.Json
 
 /**
- *  NickNameRequest.kt
+ *  NotifyResponse.kt
  *
- *  Created by Minji Jeong on 2022/07/28
+ *  Created by Minji Jeong on 2022/08/02
  *  Copyright © 2022 MashUp All rights reserved.
  */
 
-data class NickNameRequest(
+data class NotifyResponse(
     @Json(name = "nickname") val nickName: String,
+    @Json(name = "userNo") val userNo: Int,
     @Json(name = "notification") val notification: Boolean
 )
